refactor: Refactoring PostControllerTest
