+++ src/main/kotlin/me/insiro/home/server/post/exception/PostNotFoundException.kt
@@ -1,0 +1,7 @@
+package me.insiro.home.server.post.exception
+
+import me.insiro.home.server.application.exception.AbsException
+import me.insiro.home.server.post.entity.Post
+import org.springframework.http.HttpStatus
+
+class PostNotFoundException(id: Post.Id) : AbsException(HttpStatus.NOT_FOUND, "Post Not Found (id : $id )")
