+++ src/main/kotlin/me/insiro/home/server/post/service/CategoryService.kt
@@ -1,0 +1,18 @@
+package me.insiro.home.server.post.service
+
+import me.insiro.home.server.post.entity.Category
+import org.springframework.stereotype.Service
+
+@Service
+class CategoryService {
+    fun get(name: String): Category? {
+        TODO("Not Yet Implemented")
+
+    }
+
+    fun get(id: Category.Id): Category? {
+        TODO("Not Yet Implemented")
+
+    }
+
+}
