+++ src/test/kotlin/me/insiro/home/server/post/controller/PostControllerTest.kt
@@ -1,12 +1,12 @@
 package me.insiro.home.server.post.controller
 
 import me.insiro.home.server.application.domain.Status
-import me.insiro.home.server.application.exception.AbsException
-import me.insiro.home.server.post.dto.comment.CommentDTO
 import me.insiro.home.server.post.dto.post.NewPostDTO
-import me.insiro.home.server.post.dto.post.PostResponseDTO
 import me.insiro.home.server.post.dto.post.UpdatePostDTO
 import me.insiro.home.server.post.entity.*
+import me.insiro.home.server.post.service.CategoryService
+import me.insiro.home.server.post.service.CommentService
+import me.insiro.home.server.post.service.PostService
 import me.insiro.home.server.testUtils.AbsControllerTest
 import me.insiro.home.server.user.dto.AuthDetail
 import me.insiro.home.server.user.entity.User
@@ -17,18 +17,13 @@ import org.mockito.Mock
 import org.mockito.Mockito
 import org.mockito.Mockito.mock
 import org.mockito.junit.jupiter.MockitoExtension
-import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
-import org.springframework.security.access.annotation.Secured
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.stereotype.Service
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import org.springframework.web.bind.annotation.*
 
 @ExtendWith(MockitoExtension::class)
 class PostControllerTest : AbsControllerTest("/posts") {
@@ -141,124 +136,4 @@ class PostControllerTest : AbsControllerTest("/posts") {
 
 }
 
-class PostNotFoundException(id: Post.Id) : AbsException(HttpStatus.NOT_FOUND, "Post Not Found (id : $id )")
-class CategoryNotFoundException : AbsException {
-    constructor(id: Category.Id) : super(HttpStatus.NOT_FOUND, "Category Not Found (id : $id )")
-    constructor(name: String) : super(HttpStatus.NOT_FOUND, "Category Not Found (name : $name )")
-}
 
-@RestController
-@RequestMapping("posts")
-class PostController(
-    private val postService: PostService,
-    private val categoryService: CategoryService,
-    val commentService: CommentService,
-) {
-    @GetMapping
-    fun getPosts(): ResponseEntity<List<PostResponseDTO>> {
-        val posts = postService.findPosts().map { PostResponseDTO(it) }
-        return ResponseEntity(posts, HttpStatus.OK)
-    }
-
-    @Secured("ROLE_WRITER")
-    @PostMapping
-    fun createPost(
-        @RequestBody newPostDTO: NewPostDTO,
-    ): ResponseEntity<PostResponseDTO> {
-        val categoryId = newPostDTO.category?.let {
-            categoryService.get(it) ?: throw CategoryNotFoundException(it)
-        }?.id
-
-
-        val post = postService.createPost(newPostDTO, categoryId, getSignedUser().user)
-        return ResponseEntity(PostResponseDTO(post), HttpStatus.CREATED)
-    }
-
-    fun getSignedUser(): AuthDetail {
-        return SecurityContextHolder.getContext().authentication.principal as AuthDetail
-    }
-
-    @GetMapping("{id}")
-    fun getPost(
-        @PathVariable id: Post.Id,
-        @RequestParam(required = false) comment: Boolean = false
-    ): ResponseEntity<PostResponseDTO> {
-        val post = postService.findPost(id) ?: throw PostNotFoundException(id)
-        val comments = commentService.findComments(id).map { CommentDTO(it) }
-
-        return ResponseEntity(PostResponseDTO(post, comments), HttpStatus.OK)
-    }
-
-    @Secured("ROLE_WRITER")
-    @PatchMapping("{id}")
-    fun updatePost(
-        @PathVariable id: Post.Id,
-        @RequestBody updateDTO: UpdatePostDTO,
-    ): ResponseEntity<PostResponseDTO> {
-
-        val categoryId = updateDTO.category?.let {
-            categoryService.get(it) ?: throw CategoryNotFoundException(it)
-        }?.id
-        val post =
-            postService.updatePost(id, updateDTO, categoryId, getSignedUser().user) ?: throw PostNotFoundException(id)
-        return ResponseEntity(PostResponseDTO(post), HttpStatus.OK)
-    }
-
-    @DeleteMapping("{id}")
-    fun deletePost(
-        @PathVariable id: Post.Id,
-    ): ResponseEntity<Boolean> {
-        val result = postService.deletePost(id, getSignedUser().user)
-        commentService.deleteComment(id)
-        return ResponseEntity(result, HttpStatus.OK)
-    }
-}
-
-@Service
-class PostService {
-    fun createPost(createDTO: NewPostDTO, categoryId: Category.Id?, user: User): JoinedPost {
-        TODO("Not Yet Implemented")
-    }
-
-    fun updatePost(id: Post.Id, updateDTO: UpdatePostDTO, categoryId: Category.Id?, user: User): JoinedPost? {
-        TODO("Not Yet Implemented")
-    }
-
-    fun deletePost(id: Post.Id, user: User): Boolean {
-        TODO("Not Yet Implemented")
-    }
-
-    fun findPost(id: Post.Id): JoinedPost? {
-        TODO("Not Yet Implemented")
-    }
-
-    fun findPosts(): List<JoinedPost> {
-        TODO("Not Yet Implemented")
-    }
-}
-
-@Service
-class CategoryService {
-    fun get(name: String): Category? {
-        TODO("Not Yet Implemented")
-
-    }
-
-    fun get(id: Category.Id): Category? {
-        TODO("Not Yet Implemented")
-
-    }
-
-}
-
-@Service
-class CommentService {
-    fun findComments(id: Post.Id): List<Comment> {
-        TODO("Not Yet Implemented")
-    }
-
-    fun deleteComment(id: Post.Id): Int {
-        TODO("Not Yet Implemented")
-
-    }
-}
